+++ CHANGELOG.md
@@ -10,9 +10,21 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Enclosure Group
 - Ethernet Network
 - FC Network
-- FCOE Network 
+- FCOE Network
 - Hypervisor Cluster Profile
+- Hypervisor Manager
+- Interconnect
+- Interconnect Type
+- Logical Interconnect
+- Logical Interconnect Group
+- Network Set
 - Scopes
+- Server Hardware
+- Server Hardware Type
+- Storage System
+- Storage Volume
+- Storage Volume Attachment
+- Storage Volume Template
 - Uplink Set
 
 # [v1.3.0]